+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Verify bundled Monica API source freshness ([monica-api-source-freshness](https://github.com/unbraind/monica-cli/blob/master/.agents/pm/issues/monica-api-source-freshness.toon))
 - Adopt lossless official TOON and executable output contracts ([monica-official-toon-contracts](https://github.com/unbraind/monica-cli/blob/master/.agents/pm/issues/monica-official-toon-contracts.toon))
 
 ## 2026.7.24 - 2026-07-24